+++ NEWS.md
@@ -1,11 +1,13 @@
-# bfsl 0.1.0.9008
+# bfsl 0.1.0.9009
 
 * Updated README.md.
 * Using `arrows()` instead of `segments()` to plot error bars in `plot.bfsl()`.
 * Renamed sample data `pearson_york` to `pearson_york_data`.
 * Changed license to MIT.
 * Added missing "/n" in `print.bfsl()` (#1).
-* Added a `predict.bfsl()` method to calculate confidence intervals (#2).
+* Added `predict.bfsl()` method to calculate confidence intervals (#2).
+* Added `bfsl.formula()` method. This allows to plot confidence intervals
+  with `geom_smooth()` of `ggplot2`.
 
 # bfsl 0.1.0